Fixed betterResults bug, cleaned up debug statements
In computing betterResults, some results were being dropped out before being sorted; namely those with 0 ngram overlap with the search string.  So, there was a problem when there were results from ngram searching, but 0 were sorted, and the resulting size of betterResults was 0.  However, since the sorted results were size 0, no discount was created for the confidence, but also the original results were taken as good results instead of the empty betterResults.  Now, regardless of the outcome, if ngrams searching is used, betterResults is the resulting set (even if it is empty).  This should fix some false positives.

Added a debug constant in the GeonamesHelper to turn on/off debug printing.  Cleaned up the printing of debug information.