Updated confidence score cutoff
Confidence score, located in the xslt java-geo-lib.xsl file, can be tweaked for the tradeoff between false positives and false negatives.  The higher it is, the less likely items are to match.  Empirical analysis suggests that 0.06 gives a 2 percent false positive rate in my 871 item sample.  However, with this value 269 will not be matched (30 percent) compared with 230 (26 percent) with a lower value of 0.005. Many of the non-matched values are true non matches, such as subjects in the placeEntry tag.